backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
openai==1.57.2
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
pdfplumber==0.11.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
google-auth==2.29.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
